Realised rotation, validation and try to start application

+++ solution/include/rotation.h
@@ -3,6 +3,6 @@
 
 #include <image.h>
 
-struct image rotate(struct image* old, int16_t angle);
+struct image rotate(struct image* old, int64_t angle);
 
 #endif
